+++ Routes/websiteRouter.js
@@ -3,7 +3,6 @@ const websiteRout = express.Router();
 import websiteController from '../controllers/websiteController.js';
 
 websiteRout.get('/', websiteController.getall);
-websiteRout.put('/update', websiteController.update);
-websiteRout.get('/validcpu', websiteController.validCpu);
+websiteRout.put('/', websiteController.update);
 
 export default websiteRout;
